+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-04-07
+
+### Added
+
+- Add support for [Mantine v9](https://mantine.dev/changelog/9-0-0/)
+- Add `fontWeights` constants from updated theme object
+
+### Changed
+
+- Update example to use new `fontWeights` constant
+
 ## [0.1.1] - 2026-01-07
 
 ### Changed